Question REST_Api Endpoints

+++ Trello_quora/quora-service/src/main/java/com/upgrad/quora/service/business/QuestionService.java
@@ -23,16 +23,17 @@ public class QuestionService {
     @Autowired
     private UserDao userDao;
 
+    // This method is for Creating the Question
     @Transactional(propagation = Propagation.REQUIRED)
     public QuestionEntity create(QuestionEntity questionEntity, final String authorizationToken) throws AuthorizationFailedException {
         UserAuthTokenEntity userAuthTokenEntity = questionDao.getUserAuthToken(authorizationToken);
 
-        // Checking if User has Signed-In or not
+        // Validating if User has Signed-In or not
         if (userAuthTokenEntity == null) {
             throw new AuthorizationFailedException("ATHR-001", "User has not signed in");
         }
 
-        // Checking whether the user is Signed-Out or not
+        // Validating whether the user is Signed-Out or not
         if (userAuthTokenEntity.getLogoutAt() != null) {
             throw new AuthorizationFailedException("ATHR-002", "User is signed out.Sign in first to post a question");
         }
@@ -41,35 +42,47 @@ public class QuestionService {
         return questionDao.create(questionEntity);
     }
 
+    // This method is for Getting All Questions for any user
     @Transactional(propagation = Propagation.REQUIRED)
     public List<QuestionEntity> getAllQuestions(final String authorizationToken) throws AuthorizationFailedException {
         UserAuthTokenEntity userAuthTokenEntity = questionDao.getUserAuthToken(authorizationToken);
 
-        // Checking if User has Signed-In or not
+        // Validating if the User has Signed-In or not
         if (userAuthTokenEntity == null) {
             throw new AuthorizationFailedException("ATHR-001", "User has not signed in");
         }
 
-        // Checking whether the user is Signed-Out or not
-        if (userAuthTokenEntity.getLogoutAt() != null) {
-            throw new AuthorizationFailedException("ATHR-002", "User is signed out.Sign in first to get all questions");
+        // Validating whether the user is Signed-Out or not
+        else if (userAuthTokenEntity.getLogoutAt() != null) {
+            throw new AuthorizationFailedException("ATHR-002", "User is signed out. Sign in first to get all questions");
         }
 
-        List<QuestionEntity> allQuestions = questionDao.getAllQuestions();
-        return allQuestions;
+        else {
+            List<QuestionEntity> allQuestions = questionDao.getAllQuestions();
+            return allQuestions;
+        }
     }
 
+    // This method is for Editing the Question
     @Transactional(propagation = Propagation.REQUIRED)
     public QuestionEntity editQuestion(QuestionEntity questionEntity, final String questionId, final String authorizationToken) throws AuthorizationFailedException, InvalidQuestionException {
         UserAuthTokenEntity userAuthTokenEntity = questionDao.getUserAuthToken(authorizationToken);
-        // Checking if User has Signed-In or not
+
+        // Validating if User has Signed-In or not
         if (userAuthTokenEntity == null) {
             throw new AuthorizationFailedException("ATHR-001", "User has not signed in");
         }
 
-        // Checking whether the user is Signed-Out or not
+        // Validating whether the user is Signed-Out or not
         if (userAuthTokenEntity.getLogoutAt() != null) {
             throw new AuthorizationFailedException("ATHR-002", "User is signed out.Sign in first to edit the question");
+        }
+
+
+        // Validating if the Question with corresponding Uuid exist in the DataBase or  not
+        QuestionEntity question_Entity0 = questionDao.getQuestion(questionId);
+        if (question_Entity0 == null) {
+            throw new InvalidQuestionException("QUES-001", "Entered question uuid does not exist");
         }
 
         // Validating the Owner of the Question
@@ -77,18 +90,14 @@ public class QuestionService {
         if (!userEntity.getUuid().equals(userAuthTokenEntity.getUuid())) {
             throw new AuthorizationFailedException("ATHR-003", "Only the question owner can edit the question");
         }
-
-        // Validating if the Question with corresponding Uuid exist in the DataBase or  not
-        QuestionEntity question_Entity = questionDao.getQuestion(questionId);
-        if (question_Entity == null) {
-            throw new InvalidQuestionException("QUES-001", "Entered question uuid does not exist");
+        else {
+            questionEntity.setUser(userAuthTokenEntity.getUser());
+            questionDao.editQuestion(questionEntity);
+            return questionEntity;
         }
-
-        questionEntity.setUser(userAuthTokenEntity.getUser());
-        questionDao.editQuestion(questionEntity);
-        return questionEntity;
     }
 
+    // This method is for Deleting the Question
     @Transactional(propagation = Propagation.REQUIRED)
     public void deleteQuestion(final String questionId, final String authorizationToken) throws AuthorizationFailedException, InvalidQuestionException {
         UserAuthTokenEntity userAuthTokenEntity = questionDao.getUserAuthToken(authorizationToken);
@@ -103,12 +112,6 @@ public class QuestionService {
             throw new AuthorizationFailedException("ATHR-002", "User is signed out.Sign in first to delete a question");
         }
 
-        //Validating whether user is Admin or Owner
-        String role = userAuthTokenEntity.getUser().getRole();
-        UserEntity userEntity = questionDao.getQuestion(questionId).getUser();
-        if (role.equals("nonadmin") || !userEntity.getUuid().equals(userAuthTokenEntity.getUuid())) {
-            throw new AuthorizationFailedException("ATHR-003", "Only the question owner or admin can delete the question");
-        }
 
         // Validating if the Question with corresponding Uuid exist in the DataBase or  not
         QuestionEntity questionEntity = questionDao.getQuestion(questionId);
@@ -116,9 +119,18 @@ public class QuestionService {
             throw new InvalidQuestionException("QUES-001", "Entered question uuid does not exist");
         }
 
-        questionDao.deleteQuestion(questionId);
+        //Validating whether user is Admin or Owner of the question
+        String role = userAuthTokenEntity.getUser().getRole();
+        UserEntity userEntity = questionDao.getQuestion(questionId).getUser();
+        if (role.equals("admin") || userEntity.getUuid().equals(userAuthTokenEntity.getUuid())) {
+            questionDao.deleteQuestion(questionId);
+        }
+        else {
+            throw new AuthorizationFailedException("ATHR-003", "Only the question owner or admin can delete the question");
+        }
     }
 
+    // This method is for Getting All the Questions posted by respective user
     @Transactional(propagation = Propagation.REQUIRED)
     public List<QuestionEntity> getAllQuestionsByUser(final String userId, final String authorizationToken) throws AuthorizationFailedException, InvalidQuestionException, UserNotFoundException {
         UserAuthTokenEntity userAuthTokenEntity = userDao.getUserAuthToken(authorizationToken);
@@ -133,12 +145,14 @@ public class QuestionService {
             throw new AuthorizationFailedException("ATHR-002", "User is signed out.Sign in first to get all questions posted by a specific user");
         }
 
+        //Validate if user exist in the data base or not.
         UserEntity userEntity = userDao.getUser(userId);
         if (userEntity == null) {
             throw new UserNotFoundException("USR-001", "User with entered uuid whose question details are to be seen does not exist");
         }
-
-        List<QuestionEntity> allQuestions = questionDao.getAllQuestionsByUser(userId);
-        return allQuestions;
+        else {
+            List<QuestionEntity> allQuestions = questionDao.getAllQuestionsByUser(userId);
+            return allQuestions;
+        }
     }
 }
